+++ src/pages/APISelect/APISelectPage.tsx
@@ -2,29 +2,12 @@ import Button from '../../components/@common/Button/Button.styles';
 import Template from '../../components/@common/Template/Template';
 import Title from '../../components/@common/Title/Title.styles';
 import { APISelectList } from './APISelectPage.styles';
-import { currentAPI } from '../../API/API';
-import { useEffect, VFC } from 'react';
+import { VFC } from 'react';
 import { RouteComponentProps } from 'react-router';
-import { useAppDispatch } from '../../state/store';
-import { APIAction } from '../../state/slices/API';
-import { lineAction } from '../../state/slices/line';
-import { loginAction } from '../../state/slices/login';
-import { API_URL } from '../../constants/API';
-import { APIName } from '../../types';
+import useAPISelect from '../../hooks/useAPISelect';
 
-const APISelectPage: VFC<RouteComponentProps> = ({ history }) => {
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(lineAction.initLineId());
-    dispatch(loginAction.logout());
-  }, []);
-
-  const selectAPI = (name: APIName) => {
-    currentAPI.baseUrl = API_URL[name];
-    dispatch(APIAction.setName(name));
-    history.push('/login');
-  };
+const APISelectPage: VFC<RouteComponentProps> = () => {
+  const { selectAPI } = useAPISelect();
 
   return (
     <Template type="vertical">
